% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone_fun.R
\name{CreateTableOne_Strata}
\alias{CreateTableOne_Strata}
\title{CreateTableOne_Strata}
\usage{
CreateTableOne_Strata(vars, strata, data, keep_test = FALSE,
  nonnormal = NULL, ...)
}
\arguments{
\item{vars}{Variables to be summarized given as a character vector. Factors are handled as categorical variables, whereas numeric variables are handled as continuous variables. If empty, all variables in the data frame specified in the data argument are used.}

\item{strata}{Stratifying (grouping) variable name(s) given as a character vector.}

\item{data}{A data frame in which these variables exist. All variables (both vars and strata) must be in this data frame.}

\item{keep_test}{binary: keep the column "test"}

\item{nonnormal}{A character vector to specify the variables for which the p-values should be those of nonparametric tests. By default all p-values are from normal assumption-based tests (oneway.test).}

\item{...}{need to add this: other arguments to ?}
}
\value{
a tibble
}
\description{
An extension to tableone::CreateTableOne that makes a table of overall information,
cbinded to a strata table with p-values, adding a column of number of non-missing
values for each variable, returning the printed table (from print.TableOne) as a data frame
with no rownames. Column names include strata name.
}
\examples{

mtcars2 = mtcars\%>\%mutate(gear=factor(gear),vs=factor(vs))
CreateTableOne_Strata(vars=c("mpg","vs","gear"),strata="cyl",data=mtcars2)
}
